% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{Graficos}
\alias{Graficos}
\title{Funcion para generar graficos histograma y boxplot}
\usage{
Graficos(x)
}
\arguments{
\item{x}{parametro de entrada dataframe}
}
\description{
Funcion para generar graficos histograma y boxplot
}
