% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{replace.na<-}
\alias{replace.na<-}
\title{Funcion de Reemplazo}
\usage{
replace.na(x) <- value
}
\arguments{
\item{x}{primer parametro dataframe de solo variables numericas}

\item{value}{segundo parametro parametro de reemplazo}
}
\description{
Funcion de Reemplazo
}
