% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{solnum}
\alias{solnum}
\title{Funcion para extraer solo variables numericas}
\usage{
solnum(x)
}
\arguments{
\item{x}{parametro de entrada corresponde al dataframe inicial}
}
\description{
Funcion para extraer solo variables numericas
}
